+++ R/misc.func.hidden.uni.r
@@ -265,7 +265,7 @@
 
 }
 
-.rma.ls.solnp.ineqfun <- function(par, yi, vi, X, Z, reml, k, pX, alpha.val, beta.val, verbose, digits, REMLf, link, mZ, alpha.min, alpha.max, alpha.transf, tau2.min, tau2.max, optbeta) {
+.rma.ls.ineqfun.pos <- function(par, yi, vi, X, Z, reml, k, pX, alpha.val, beta.val, verbose, digits, REMLf, link, mZ, alpha.min, alpha.max, alpha.transf, tau2.min, tau2.max, optbeta) {
 
    if (optbeta) {
       alpha <- par[-seq_len(pX)]
@@ -279,6 +279,25 @@
    alpha <- ifelse(is.na(alpha.val), alpha, alpha.val)
 
    tau2 <- c(Z %*% alpha)
+
+   return(tau2)
+
+}
+
+.rma.ls.ineqfun.neg <- function(par, yi, vi, X, Z, reml, k, pX, alpha.val, beta.val, verbose, digits, REMLf, link, mZ, alpha.min, alpha.max, alpha.transf, tau2.min, tau2.max, optbeta) {
+
+   if (optbeta) {
+      alpha <- par[-seq_len(pX)]
+   } else {
+      alpha <- par
+   }
+
+   if (alpha.transf)
+      alpha <- mapply(.mapfun.alpha, alpha, alpha.min, alpha.max)
+
+   alpha <- ifelse(is.na(alpha.val), alpha, alpha.val)
+
+   tau2 <- -c(Z %*% alpha)
 
    return(tau2)